+++ src/file/ini_file_spider.py
@@ -107,6 +107,7 @@ def write_minio_config_to_file(minio_config):
     conf.set("automatic_config", "allow_replace_domain_flag", 'True')
     conf.set("automatic_config", "scheduled_download_program_flag", 'True')
     conf.set("automatic_config", "dmi_api_server", '192.168.163.129:8888')
+    conf.set("automatic_config", "detect_img_model", "python")
 
     conf.add_section("minio_config_selected")
     conf.set("minio_config_selected", "minio_config_id", "1")
@@ -208,6 +209,7 @@ def check_ini_config():
         conf.set("automatic_config", "scheduled_download_program_flag", 'True')
         # dmi_api_server
         conf.set("automatic_config", "dmi_api_server", '192.168.163.129:8888')
+        conf.set("automatic_config", "detect_img_model", "python")
 
         conf.add_section("minio_config_selected")
         conf.set("minio_config_selected", "minio_config_id", "1")